requirements.txt
alabaster==0.7.10
appdirs==1.4.3
asgiref==2.2.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
autopep8==1.4.2
Babel==2.5.3
certifi==2018.1.18
channels==2.0.2
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
daphne==2.1.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.13
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-filter==2.1.0
django-guardian==1.4.9
django-haystack==2.8.0
django-js-asset==1.0.0
django-mailbox==4.7.1
django-money==0.12.3
django-mptt==0.9.0
django-nose==1.4.5
django-polymorphic==2.0.2
django-polymorphic-tree==1.5
django-rest-swagger==2.1.2
django-silk==3.0.1
django-sphinxdoc==1.5.1
django-tag-parser==3.1
django-timezone-field==2.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
freezegun==0.3.10
future==0.16.0
google-api-python-client==1.6.5
googlemaps==2.5.1
gprof2dot==2016.10.13
-e git+git@bitbucket.org:fullplanner/gth-core.git@29f4f7ea99850498919c13644ab84b5f061fdd2e#egg=gth_core
-e git+https://godardth:maCE12ku@bitbucket.org/fullplanner/gth-project@c830bf86aa2a9e6c2021613fe6f02332561a4c4a#egg=gth_project
gunicorn==19.7.1
httplib2==0.10.3
hyperlink==18.0.0
idna==2.6
imagesize==1.0.0
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
jsonpickle==0.9.6
lpthw.web==1.1
lxml==4.2.0
Markdown==3.0.1
MarkupSafe==1.0
nose==1.3.7
nose-exclude==0.5.0
oauth2client==4.1.2
oauthlib==2.0.6
openapi-codec==1.3.2
packaging==17.1
pathlib==1.0.1
Pillow==5.4.1
pipdeptree==0.11.0
psycopg2-binary==2.7.4
pudb==2017.1.4
py-moneyed==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.4.0
pycryptodomex==3.5.1
Pygments==2.2.0
pyjwkest==1.4.0
PyJWT==1.6.0
pyparsing==2.2.0
python-dateutil==2.7.0
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
rest-social-auth==1.2.0
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
snowballstemmer==1.2.1
social-auth-app-django==1.2.0
social-auth-core==1.7.0
spec==1.4.1
Sphinx==1.7.1
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
Twisted==17.9.0
txaio==2.9.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
urwid==2.0.1
virtualenv==15.1.0
zope.interface==4.4.3
